+++ src/components/Indexer/processors/index.ts
@@ -1,0 +1,16 @@
+import { BaseEventProcessor } from './BaseProcessor'
+
+export * from './DispenserActivatedEventProcessor.js'
+export * from './DispenserCreatedEventProcessor.js'
+export * from './DispenserDeactivatedEventProcessor.js'
+export * from './ExchangeActivatedEventProcessor.js'
+export * from './ExchangeCreatedEventProcessor.js'
+export * from './ExchangeDeactivatedEventProcessor.js'
+export * from './ExchangeRateChangedEventProcessor.js'
+export * from './MetadataEventProcessor.js'
+export * from './MetadataStateEventProcessor.js'
+export * from './OrderReusedEventProcessor.js'
+export * from './OrderStartedEventProcessor.js'
+export * from './BaseProcessor.js'
+
+export type ProcessorConstructor = new (chainId: number) => BaseEventProcessor
